finished splitting code
