+++ css/custom/styles.css
@@ -19,9 +19,13 @@ h3 {
   font-weight: 400;
 }
 
+h5 {
+  color: #6118fb;
+}
+
 a {
   text-decoration: none;
-  color: #6118fb;
+  color: #d4bfff;
 }
 a:hover {
   color: #fff;
