+++ home.html
@@ -17,7 +17,7 @@
         <Header>
             <nav class="navbar navbar-expand-md navbar-dark fixed-top bg-black">
                 <div class="container-fluid">
-                    <a class="navbar-brand" href="index.html"><img src="images/logo/Vectorts_logo_light.png" alt=""></a>
+                    <a class="navbar-brand" href="home.html"><img src="images/logo/Vectorts_logo_light.png" alt=""></a>
                     <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
                         data-bs-target="#navbarCollapse" aria-controls="navbarCollapse" aria-expanded="false"
                         aria-label="Toggle navigation">
