+++ js/modules/functions/posts/getAllPosts.mjs
@@ -38,12 +38,12 @@ export async function fetchPosts(url) {
                 <div class="topCard">
                     <h4 class="card-title">${title}</h4>
                     <div class="updateEdit">
-                    <a href="update.html?id=${id}" class="updateLink"><h5>Edit or Delete Post</h5></a>
+                   <h5>Edit or Delete Post</h5>
                      </div>
                     
                     </div>
                     <p class="card-text">${body}</p>
-                    <p class="card-text"><small class="text-muted created">Created: ${postDate} at ${postTime} by <b>${postAuthor}</b></small></p>
+                    <p class="card-text"><small class="text-muted created">Created: ${postDate} at ${postTime} by <b>${postAuthor}</b> - post: ${id}</small></p>
                 
                 </div>
             </div>
@@ -58,7 +58,7 @@ export async function fetchPosts(url) {
                     <h4 class="card-title">${title}</h4>
                      <div class="updateEdit"> </div>
                     <p class="card-text">${body}</p>
-                    <p class="card-text"><small class="text-muted created">Created: ${postDate} at ${postTime} by <b>${postAuthor}</b></small></p>
+                    <p class="card-text"><small class="text-muted created">Created: ${postDate} at ${postTime} by <b>${postAuthor}</b> - post: ${id}</small></p>
                 
                 </div>
             </div>
