+++ js/modules/functions/user/isSignedIn.mjs
@@ -1,0 +1,9 @@
+import { token, userName } from "../../variables/variables.mjs";
+
+export function checkIfsignedIn() {
+  if (token && userName) {
+    window.location.href = `home.html`;
+  }
+}
+
+// checkIfsignedIn();
