+++ js/modules/functions/user/signout.mjs
@@ -2,6 +2,7 @@
 
 export function signOut() {
   localStorage.removeItem("accessToken");
+  localStorage.removeItem("userName");
 }
 
 // signoutBtn.addEventListener("click", signOut);
