+++ js/register.mjs
@@ -1,0 +1,7 @@
+import { regForm } from "./modules/variables/variables.mjs";
+import { register } from "./modules/functions/user/create.mjs";
+import { checkIfsignedIn } from "./modules/functions/user/isSignedIn.mjs";
+
+regForm.addEventListener("submit", register);
+
+checkIfsignedIn();
